+++ CHANGELOG.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Change type of emoji reaction notification to `pleroma:emoji_reaction`.
 - Set limit on multipart form size.
 
+### Removed
+
+- Remove `sameAs` property from portable actor objects.
+
 ### Fixed
 
 - Prevent self-follow when moving followers to another account.